+++ BUILD
@@ -504,8 +504,7 @@ grpc_cc_library(
         "src/core/ext/transport/binder/client/endpoint_binder_pool.cc",
         "src/core/ext/transport/binder/client/jni_utils.cc",
         "src/core/ext/transport/binder/client/security_policy_setting.cc",
-        "src/core/ext/transport/binder/security_policy/internal_only_security_policy.cc",
-        "src/core/ext/transport/binder/security_policy/untrusted_security_policy.cc",
+        "src/core/ext/transport/binder/security_policy/binder_security_policy.cc",
         "src/core/ext/transport/binder/server/binder_server.cc",
         "src/core/ext/transport/binder/server/binder_server_credentials.cc",
         "src/core/ext/transport/binder/transport/binder_transport.cc",
@@ -518,17 +517,12 @@ grpc_cc_library(
     ],
     hdrs = [
         "src/core/ext/transport/binder/client/binder_connector.h",
-        "src/core/ext/transport/binder/client/channel_create.h",
         "src/core/ext/transport/binder/client/channel_create_impl.h",
         "src/core/ext/transport/binder/client/connection_id_generator.h",
         "src/core/ext/transport/binder/client/endpoint_binder_pool.h",
         "src/core/ext/transport/binder/client/jni_utils.h",
         "src/core/ext/transport/binder/client/security_policy_setting.h",
-        "src/core/ext/transport/binder/security_policy/internal_only_security_policy.h",
-        "src/core/ext/transport/binder/security_policy/security_policy.h",
-        "src/core/ext/transport/binder/security_policy/untrusted_security_policy.h",
         "src/core/ext/transport/binder/server/binder_server.h",
-        "src/core/ext/transport/binder/server/binder_server_credentials.h",
         "src/core/ext/transport/binder/transport/binder_stream.h",
         "src/core/ext/transport/binder/transport/binder_transport.h",
         "src/core/ext/transport/binder/utils/transport_stream_receiver.h",
@@ -552,16 +546,18 @@ grpc_cc_library(
         "absl/time",
     ],
     language = "c++",
-    # TODO(mingcl): Move public headers under include/ and put them here
     public_hdrs = [
+        "include/grpcpp/security/binder_security_policy.h",
+        "include/grpcpp/create_channel_binder.h",
+        "include/grpcpp/security/binder_credentials.h",
     ],
+    visibility = ["@grpc:public"],
     deps = [
         "gpr",
         "gpr_base",
         "gpr_platform",
         "grpc",
         "grpc++_base",
-        "grpc++_internals",
         "grpc_base",
         "grpc_client_channel",
         "grpc_codegen",
