+++ examples/android/binder/java/io/grpc/binder/cpp/exampleclient/native.cc
@@ -18,8 +18,8 @@
 #include "examples/protos/helloworld.grpc.pb.h"
 #include "examples/protos/helloworld.pb.h"
 
-#include "src/core/ext/transport/binder/client/channel_create.h"
-#include "src/core/ext/transport/binder/security_policy/untrusted_security_policy.h"
+#include <grpcpp/create_channel_binder.h>
+#include <grpcpp/security/binder_security_policy.h>
 
 extern "C" JNIEXPORT jstring JNICALL
 Java_io_grpc_binder_cpp_exampleclient_ButtonPressHandler_native_1entry(
