+++ examples/android/binder/java/io/grpc/binder/cpp/exampleserver/native.cc
@@ -23,10 +23,8 @@
 #include "examples/protos/helloworld.pb.h"
 
 #include <grpcpp/grpcpp.h>
-
-#include "src/core/ext/transport/binder/security_policy/untrusted_security_policy.h"
-#include "src/core/ext/transport/binder/server/binder_server.h"
-#include "src/core/ext/transport/binder/server/binder_server_credentials.h"
+#include <grpcpp/security/binder_credentials.h>
+#include <grpcpp/security/binder_security_policy.h>
 
 namespace {
 class GreeterService : public helloworld::Greeter::Service {
