+++ src/core/ext/transport/binder/client/channel_create_impl.h
@@ -17,7 +17,8 @@
 
 #include <grpc/support/port_platform.h>
 
-#include "src/core/ext/transport/binder/security_policy/security_policy.h"
+#include <grpcpp/security/binder_security_policy.h>
+
 #include "src/core/ext/transport/binder/wire_format/binder.h"
 #include "src/core/lib/channel/channel_args.h"
