+++ src/core/ext/transport/binder/client/security_policy_setting.h
@@ -20,7 +20,8 @@
 #include "absl/container/flat_hash_map.h"
 #include "absl/strings/string_view.h"
 
-#include "src/core/ext/transport/binder/security_policy/security_policy.h"
+#include <grpcpp/security/binder_security_policy.h>
+
 #include "src/core/lib/gprpp/sync.h"
 
 namespace grpc_binder {
