+++ src/core/ext/transport/binder/server/binder_server.h
@@ -22,7 +22,8 @@
 #include "absl/container/flat_hash_map.h"
 #include "absl/status/status.h"
 
-#include "src/core/ext/transport/binder/security_policy/security_policy.h"
+#include <grpcpp/security/binder_security_policy.h>
+
 #include "src/core/ext/transport/binder/transport/binder_transport.h"
 #include "src/core/ext/transport/binder/wire_format/binder.h"
 #include "src/core/lib/iomgr/exec_ctx.h"
