+++ src/core/ext/transport/binder/server/binder_server_credentials.cc
@@ -14,9 +14,9 @@
 
 #include <grpc/support/port_platform.h>
 
+#include <grpcpp/security/binder_security_policy.h>
 #include <grpcpp/security/server_credentials.h>
 
-#include "src/core/ext/transport/binder/security_policy/security_policy.h"
 #include "src/core/ext/transport/binder/server/binder_server.h"
 #include "src/core/ext/transport/binder/wire_format/binder_android.h"
