+++ test/core/transport/binder/end2end/binder_server_test.cc
@@ -24,11 +24,11 @@
 
 #include <grpcpp/grpcpp.h>
 #include <grpcpp/impl/grpc_library.h>
+#include <grpcpp/security/binder_credentials.h>
+#include <grpcpp/security/binder_security_policy.h>
 
 #include "src/core/ext/transport/binder/client/channel_create_impl.h"
-#include "src/core/ext/transport/binder/security_policy/untrusted_security_policy.h"
 #include "src/core/ext/transport/binder/server/binder_server.h"
-#include "src/core/ext/transport/binder/server/binder_server_credentials.h"
 #include "test/core/transport/binder/end2end/fake_binder.h"
 #include "test/core/util/test_config.h"
 #include "test/cpp/end2end/test_service_impl.h"
